Renamed files and removed old injection handling
